+++ frontend/src/pages/chatRoom/ChatRoom.tsx
@@ -25,7 +25,7 @@ import ChatRoomInputArea from './components/ChatRoomInputArea/ChatRoomInputArea'
 import ImageSendModal from './components/ImageSendModal/ImageSendModal';
 import MentoringActionPanel from './components/MentoringActionPanel/MentoringActionPanel';
 import { MESSAGE_TYPE } from './constants/message';
-import useDelayedVisibility from './hooks/useDelayedVisibility';
+import useDelayedVisibility from '../../common/hooks/useDelayedVisibility';
 import useImageFile from './hooks/useImageFile';
 import useInfiniteChatRoomMessage from './hooks/useInfiniteChatRoomMessage';
 import usePersistPendingMessages, {
